+++ .gitignore
@@ -52,8 +52,8 @@ coverage.xml
 .pytest_cache/
 cover/
 
-# Data
-/src/data
+# Ignore all csv files in the src/data directory
+src/data/**/*.csv
 
 # Translations
 *.mo
@@ -168,8 +168,5 @@ cython_debug/
 #  and can be added to the global gitignore or merged into this file.  For a more nuclear
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 .idea/
-
-
-temp_train_models/
 
 scripts/notebooks/mlruns/
