+++ src/data/gluroo/gluroo_data_cleaner.py
@@ -1,0 +1,7 @@
+"""Data cleaning functions for the GLUROO dataset."""
+
+import pandas as pd
+
+
+def clean_gluroo_data(data: pd.DataFrame) -> pd.DataFrame:
+    pass
