+++ src/data/gluroo/gluroo_data_loader.py
@@ -1,0 +1,7 @@
+import pandas as pd
+
+
+def _clean_gluroo_train_data(data: pd.DataFrame) -> pd.DataFrame:
+    """
+    Cleans the GLUROO dataset.
+    """
